# backend-marketing/openapi.yaml
openapi: 3.0.3
info:
  title: FeedPlug API
  description: API Backend for FeedPlug Marketing Platform
  version: 1.0.0
  contact:
    name: FeedPlug Support
servers:
  - url: /api/v1
    description: Current API version
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
        '429':
          description: Too many attempts
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                company:
                  type: string
      responses:
        '201':
          description: Registration successful
        '400':
          description: Validation error
        '409':
          description: Email already exists
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New tokens issued
        '401':
          description: Invalid refresh token
...
  /ingestion/items/{id}/enrichment-analysis:
    get:
      tags: [Enrichment]
      summary: Get enrichment analysis for item
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Item ID, UUID, MPN or SKU
      responses:
        '200':
          description: Enrichment analysis
          content:
            application/json:
              schema:
                type: object
                properties:
                  itemId:
                    type: string
                  enrichments:
                    type: object
                  alerts:
                    type: array
                    items:
                      type: string
  /enrichment/score/{itemId}:
    get:
      tags: [Enrichment]
      summary: Get quality score for item
      security:
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Quality score breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  globalScore:
                    type: number
                  breakdown:
                    type: object
                    properties:
                      title:
                        type: object
                      description:
                        type: object
                      images:
                        type: object
                      technical:
                        type: object
                  recommendations:
                    type: array
                    items:
                      type: object
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        accountId:
          type: string
        accountName:
          type: string
        plan:
          type: string
...